+++ src/components/[guild]/EditGuild/EditGuildDrawer.tsx
@@ -42,7 +42,7 @@ import { Chain } from "wagmiConfig/chains"
 import LeaveButton from "../LeaveButton"
 import useGuildPermission from "../hooks/useGuildPermission"
 import useUser from "../hooks/useUser"
-import Admins from "./components/Admins"
+import Admins from "./components/Admins/Admins"
 import BackgroundImageUploader from "./components/BackgroundImageUploader"
 import ChangingGuildPinDesignAlert from "./components/ChangingGuildPinDesignAlert"
 import ColorPicker from "./components/ColorPicker"
