--- src/components/[guild]/EditGuild/components/Admins/index.ts
@@ -1,3 +1,0 @@
-import Admins from "./Admins"
-
-export default Admins
