--- src/components/[guild]/EditGuild/components/TransferOwnership/index.ts
@@ -1,3 +1,0 @@
-import TransferOwnership from "./TransferOwnership"
-
-export default TransferOwnership
