+++ src/pages/[guild]/forms/[formId]/index.tsx
@@ -1,6 +1,6 @@
 import { Box, Icon } from "@chakra-ui/react"
 import { Question } from "@phosphor-icons/react"
-import { EditGuildDrawerProvider } from "components/[guild]/EditGuild/EditGuildDrawerContext"
+import { EditGuildDrawerProvider } from "components/[guild]/EditGuild/EditGuildDrawerProvider"
 import JoinModalProvider from "components/[guild]/JoinModal/JoinModalProvider"
 import useActiveMembershipUpdate from "components/[guild]/JoinModal/hooks/useActiveMembershipUpdate"
 import RoleRequirements from "components/[guild]/Requirements"
